Explicitly use project() statement
Instead of relying on find_package(Slicer ...) to both include
the configure prerequisites and invoke the project command, this commit
switches to a more explicit convention.

The explicit convention is the documented in [1]. It also allows
the new extension wizard to be used. [2]

Finally, we also pass the variable Slicer_PREREQUISITES_CONFIGURED
because all required variables are already passed and we want to
skip the test complaining that 'find_package(Slicer COMPONENTS ConfigurePrerequisites)'
is not called before 'project()' statement.

[1] http://wiki.slicer.org/slicerWiki/index.php/Documentation/Nightly/Developers/Build_Module
[2] http://wiki.slicer.org/slicerWiki/index.php/Documentation/Nightly/Developers/ExtensionWizard#Contributing_Extensions