Centerline metrics module: Add cross-section area measurements and browsing along the centerline
Add computation of actual cross-section of a a lumen segment (or model) along a centerline curve (or model).
The lumen surface is plane can be cut at an arbitrary point of the centerline.
Intersection line segments are triangulated and the polygon closest to the centerline is kept and used for computing the cross-section area.

Cross-section and maximum inscribed sphere can be displayed for visual verification.
Display 'Diameter (MIS)' and 'Diameter (CE)' (from true cross-section area).

As a UI enhancement, a push button allows toggling between the previous and the plot layout.

Include squashed and cleaned-up commits submitted in https://github.com/vmtk/SlicerExtension-VMTK/pull/38.

Co-authored-by: Andras Lasso <lasso@queensu.ca>