Centerline metrics: show cross-section along centerline as a model and measure its area.
This patch shows an exact-fit cross-section of a segment along a centerline in a 3D view, as makes its surface area available.

It works on a closed representation of a segment, which is plane cut at an arbitrary point of the centerline. Intersection line segments are triangulated and the polygon closest to the centerline is kept and used for computing the cross-section area.

Optionally a stack of cross-sections can be created so that the polygons that are used for computing the cross-section area can be visually verified. A common scalar value (centerline point index) is assigned to each point of the cross-sections in the stack. Thresholding the scalars may then control the visibility range of cross-sections in the stack.

In addition, the diameter of a virtual circle having this area is also shown, as well as the difference to the VMTK computed diameter. (This may help to better understand some arterial stenosis.)

As a UI enhancement, a push button allows toggling between the previous and the plot layout.

N.B : a volume node is not needed.