Pass in centerline polydata instead of the MRML nodes.
This allows to use a single class for both centerline models and curves,
and a single method to determine cross-sections. This method comes from the
new one used in the Python implementation, featuring vtkParallelTransportFrame.
This also eliminates discrepencies in diameter and surface area values
found with centerline models. It was using the previous interlopating method.