Fix callback possibly missed in logic.
Previously, if the scene was closed while the module is shown, the callback
function in logic would be reset to None. On selecting a newly created curve,
the callback was not assigned again. This would result in a permanent empty
output table.

Fix by letting the widget class assigning all variables in logic.