Integrate a markups Shape::Tube node as arterial wall surface for cross-section analysis.
The wall surface provides its axial spline as centerline. All facilities
of 'Cross-section analysis' module can use this centerline.

As a side effect, by comparing the cross-section area of the lumen segment
and that of the wall, 2D stenosis degree can be calculated along the centerline.
-------------------------------------------------------------------------------

Other significant changes :

*Output table columns are created only when appropriate :
- distance : always
- MIS diameter : if present in input centerline
- CE diameter : if an input segment is selected
- Cross-section area : if an input segment is selected
- Wall diameter : if input centerline is a Shape node
- Wall surface area : if input centerline is a shape node
- Coordinates : always
List of available plots is updated altogether.This method allows to remove a
workaround used when input centerline doesn't have a radius scalar.

*Output nodes are always created, even if the input centerline does not have a
radius scalar.

*'Maximum inscribed sphere' model is hidden rather than deleted. This prevents a
crash if it is selected as the input centerline.

*Use pre-computed values from the output table only.