Miscellaneous changes.
** StenosisMeasurement3D
* Update module texts.
Help text and contributor list.

* Improve documentation.
Decribe a few ways of drawing a Shape::Tube node to represent an
artery's wall.

** QuickArterySegmentation
* Avoid cdash test failure.
The module generic tests on cdash fail in the widget's setup() on calls to
self.installExtensionFromServer().
Check slicer.app.testingEnabled() before installing an extension.

* Improve pre-fit ROI bounds.
Very often, 2 fiducial points are placed in the same slice view.
The bounds of the fiducial node will then be the same on one axis.
The ROI bounds was usually set to the exact bounds of the fiducial node, so that
it was not possible to resize the ROI in UI.
Improve ROI resizing by slightly altering the bounds in such situatons.
Simultaneously, improve UI labeling of this function.

* Improve 'Intensity tolerance' widget.
Expand tooltip and increase maximum value.

* Expand contributor list.

** GuidedArterySegmentation
* Improve 'Intensity tolerance' widget.
Expand tooltip and increase maximum value.

* Improve the use of 'Split volume' effect.
Blank out unwanted regions with the lowest intensity of the input
volume, rather than the previous default value of -1000.

* Expand contributor list.