Handle holes in lumen. (#119)
* Browse connected regions of a lumen surface.

This is intended to manage holes in the lumen surface of a segmentation
node. Model nodes are out of scope here.

The lumen surface should ideally be a single region. If multiple
regions are identified, it may mean the presence of small defects inside
the lumen, or surrounding portions of the segment that are not
contiguous to the lumen.

Each region can be focused in the slice views. After positioning the slice
views on a region, the 'Paint' effect of the 'Segment editor' can be activated
in-situ with a sphere brush to fill small defects.