Add EditCenterline module.
This module pre-fits a Shape::Tube markups node along
 - an arbitrary markups curve,
 - a centerline model,
 - a centerline curve.

An optional output centerline model or curve node can be created. It
corresponds to the spline of the created tube and has a 'Radius' array. Each
output centerline may be selectively updated when the tube is modified.

The Tube's radii can be scaled by a percent offset.