Reset output when input is modified.
CrossSectionAnalysis

Previously, the visible output was not invalidated when an input centerline,
segmentation or segment was changed in the UI widgets. It was cleared when an
output table node or plot series node was changed. The output is now cleared in
this situation, the logic being left untouched, though its content is invalid.
The changes must be applied again.