Avoid division by zero and improve logging of empty cross-sections.
CrossSectionAnalysis

When using a Tube node, the cross-section of the tube itself may be empty
sections. Stenosis calculations would then fail on a division by zero.
Increasing the resolution of the tube may resolve this.

At the same time, show the name of the surface in console and in status bar.