Dump aggregate volumes in a database.
StenosisMeasurement3D

An SQLite database is created to store aggregate volumes of the whole enclosed
lumen and of the tube. These indexed tables are created:

 - CumulativeVolumes: cumulates measurements from the first spline point to
 the last with a unit spline point step
 - BoundVolumes: provide measurements within spline point bounds, from point id
 'p' to 'p+n'.

There may be marginal differences with the results in the UI, mainly with the
lumen volume. These are inversely proportional to the spline resolution. The
surface resolution influences less.

This database is provided for further analysis in specialised statistics
software.

Update