Add a display node for the lesion model.
StenosisMeasurement3D

The sole purpose of this node is to allow customising the display of the lesion
model from the application startup file. A default node of this type with all
display preferences must be added to the scene for any effect. The properties
of the template node are applied once only to preserve any further change in
the UI.