Use the enclosed polydata in the lumen as it is clipped.
CrossSectionAnalysis

Artifacts often exist at the ends of the clipped polydata. These were previously
removed by cutting the clipped lumen at each end of the tube. This is not
friendly to archs and loops.

Use the clipped lumen as it is and inform about artifacts to be manually
removed if necessary.