Update handling of the cross-section model
CrossSectionAnalysis

The functions creating, updating and deleting the cross-section model node are
moved from logic to widget, where this is deemed more appropriate. Since this
node is meant for visualisation only and is not a real output node, it is not
saved with the scene so that another node is not recreated when working with
a saved scene.