Update handling of the maximum inscribed sphere model.
CrossSectionAnalysis

The functions creating, updating and hiding the maximum inscribed sphere model
node are moved from logic to widget, where this is deemed more appropriate.
Since this node is meant for visualisation only and is not a real output node,
it is not saved with the scene so that another node is not recreated when
working with a saved scene. This node is now removed from scene instead of
being hidden.