Show a cross-section model of a tube.
CrossSectionAnalysis

If a Shape::Tube node is used as an input centerline, a cross-section model of
the tube can be here optionally created. This may improve the visual impression
of a tight stenosis in particular, when shown together with the cross-section
model of the lumen.

At the same time, improve a few UI strings and the documentation page.