Replace an input segment by its largest region.
StenosisMeasurement3D
QuickArterySegmentation
GuidedArterySegmentation
CrossSectionAnalysis

The number of connected regions of a segment is displayed and the segment may
be replaced by the largest region. This is intended to remove holes inside a
lumen segment. The presence of even tiny holes is one of the many reasons for a
failure in vtkBooleanOperationPolyDataFilter.

Formerly, a hole removal function was implemented in QuickArterySegmentation and
it was imported in GuidedArterySegmentation and CrossSectionAnalysis. These 3
python modules now import from StenosisMeasurement3D.