Allow optional decimation of the wall polydata.
StenosisMeasurement3D
CrossSectionAnalysis

The purpose of decimating a polydata here is twofold:
 - minimise the causes of failure in vtkBooleanOperationPolyDataFilter
 - minimise the creation of empty cross-sections.

vtkBooleanOperationPolyDataFilter is used to clip the lumen within the tube in
StenosisMeasurement3D, and to clip a lumen model within the tube in
CrossSectionAnalysis. (A lumen segment in the latter module is clipped within the
tube using the segment editor.)

vtkQuadricDecimation is used and its main parameters can be tuned (API only).
The default parameters target minimal decimation.

For cross-sections of the Tube in CrossSectionAnalysis, decimation is tuned on
the higher end of the target reduction ratio to obtain a valid cross-section
polydata (not empty) on the whole spline.

Decimation is always optional. It should be used if any processing fails (lumen
clipping, cross-section). Even then, success is not guaranteed. The processing
may even succeed without any decimation and fails if any pre-processing is
applied.

At the same time:
 - rename some identifiers for clarity in CrossSectionAnalysis mainly,
 - make vtkCrossSectionCompute work on input polydata objects only, instead of
   MRML nodes,
 - restore the display node properties in
   vtkSlicerStenosisMeasurement3DLogic::UpdateSegmentBySmoothClosing(),
 - decrease logging verbosity in CrossSectionAnalysis,
 - decrease the minimal width of the CrossSectionAnalysis widget if there is no
   lumen input,
 - clear the tooltip of the result labels,
 - update inline comments.