Add a 'smooth closing' option during segmentation.
QuickArterySegmentation
GuidedArterySegmentation

The problem: the presence of holes in a segment of study.

This is detrimental to cross-section analysis, to volume analysis and to the
radius array of extracted centerlines. Holes can appear when a segment is
created using 'Flood filling' in particular.

Previously, the largest connected polydata region was converted to a segment to
eliminate these holes. The side effect is a small shrinking of the resulting
segment in 'some' areas, namely when a lumen is small. A tight stenosis may thus
get tighter.

The current approach is to fill the holes using the Smoothing effect of the
segment editor. The geometry of the lumen is better preserved without the above
mentioned side effect. But close branches may fuse together, this should be
handled on a per-case basis.