Show a wall cross-section in full or subtract the lumen.
CrossSectionAnalysis

If a tube is used to represent the arterial wall, its cross-section can be shown
as a model while browsing along the centerline (tube spline).

Previously, the lumen was forcibly subtracted from the wall cross-section. The
lumen cross-section can be shown independently.

A full wall cross-section model is now shown by default to be consistent with
the displayed surface area. The lumen-cross-section can be optionally
subtracted from the wall cross-section via a checkable menu item.