Use an adequate lumen extraction mode when creating a cross-section.
CrossSectionAnalysis

If a Tube is used together with a clipped lumen (segment or model), create the
cross-section of the lumen using the AllRegions option. In all other cases, use
the default ClosestPoint option (closest to the spline of the Tube).

This may avoid the case of the cross-section of a small hump in a very irregular
stenosis being selected, ignoring the nearby circulating lumen.