Always clear the cross-section polydata caches.
CrossSectionAnalysis

Previously, when the wall cross-section type was changed (subtract the lumen or
not), only the cross-section at this point index of the spline was updated.

By clearing the caches completely, the displayed cross-section of the wall at
any point becomes predictable when the display type is toggled many times while
browsing along the spline.