Improve the displayed colour of the cut model.
StenosisMeasurement2D

Previously, the displayed colour of the cut model for each segment followed the
scalar value of each region. If 'Limit to closest island' is true, the cut
models of multiple segments are all displayed in the same colour.

Disable scalar visibility if 'Limit to closest island' is true so as to better
distinguish between the cut models of multiple segments. Their colour follows
that of the source segment.